File Manager - Fix height jitter on selection (T946521)
