+++ scss/widgets/generic/fileManager/_sizes.scss
@@ -1,10 +1,12 @@
 @use "../sizes" as *;
 
 // adduse
+@use "../toolbar/sizes" as *;
 
 $filemanager-toolbar-separator-height: null !default;
 $filemanager-toolbar-viewmode-width: null !default;
 $filemanager-toolbar-viewmode-padding: null !default;
+$filemanager-toolbar-top-bottom-padding: null !default;
 $filemanager-breadcrumbs-item-padding-top: null !default;
 $filemanager-breadcrumbs-item-padding-bottom: null !default;
 $filemanager-details-view-checkbox-column-width: null !default;
@@ -21,6 +23,7 @@ $filemanager-progress-panel-padding-top: 6px !default;
   $filemanager-toolbar-separator-height: 36px !default;
   $filemanager-toolbar-viewmode-width: 150px !default;
   $filemanager-toolbar-viewmode-padding: 8px !default;
+  $filemanager-toolbar-top-bottom-padding: $generic-toolbar-height;
   $filemanager-breadcrumbs-item-padding-top: 5px !default;
   $filemanager-breadcrumbs-item-padding-bottom: 8px !default;
   $filemanager-details-view-checkbox-column-width: 43px !default;
@@ -36,6 +39,7 @@ $filemanager-progress-panel-padding-top: 6px !default;
   $filemanager-toolbar-separator-height: 26px !default;
   $filemanager-toolbar-viewmode-width: 130px !default;
   $filemanager-toolbar-viewmode-padding: 5px !default;
+  $filemanager-toolbar-top-bottom-padding: $generic-mobile-toolbar-height;
   $filemanager-breadcrumbs-item-padding-top: 3px !default;
   $filemanager-breadcrumbs-item-padding-bottom: 7px !default;
   $filemanager-details-view-checkbox-column-width: 39px !default;
