+++ scss/widgets/material/fileManager/_index.scss
@@ -46,6 +46,7 @@
 
   .dx-filemanager-toolbar {
     background: $base-bg;
+    min-height: $filemanager-toolbar-top-bottom-padding + 12px;
 
     &.dx-filemanager-file-toolbar,
     &.dx-filemanager-file-toolbar .dx-toolbar {
