+++ testing/tests/DevExpress.ui.widgets/fileManager.tests.js
@@ -14,6 +14,7 @@ import './fileManagerParts/thumbnailsView.tests.js';
 import './fileManagerParts/toolbar.tests.js';
 import './fileManagerParts/navigation.tests.js';
 import './fileManagerParts/selection.tests.js';
+import './fileManagerParts/scroll.tests.js';
 import './fileManagerParts/editing.tests.js';
 import './fileManagerParts/adaptivity.tests.js';
 import './fileManagerParts/editingProgress.tests.js';
